h264nal: fix bug on bitstream parser state
We were reading the `pps_id` from `pps->seq_parameter_set_id`, which
caused it the PPS to be stored under the `sps_id` ID instead of the
`pps_id`. This is normally not a problem as all the `sps_id` and
`pps_id` values are 0, but we added a SVC test which uses 2x PPS
values (0 and 1).

Tested:

Will add a unittest (SVC-based) later.