h264nal: integrate old webrtc code
We use a fork of the (old) WebRTC BitBuffer code in order to access
to bits/bytes in a binary buffer. We kept it in a separate directory
to allow easy binary replacement with the original BitBuffer.

Both BitBuffer versions have evolved since:
* WebRTC replaced BitBuffer with BitstreamReader (see [1])
* We added a new function to do byte copy in BitBuffer.

We want to remove the assumption that the BitBuffer in thie repo is in
anyway interchangeable with the (old) WebRTC BitBuffer. So we are
integrating it in the main code.

Reference:
[1] https://webrtc-review.googlesource.com/c/src/+/232125

Tested:

Build system.

```
$ rm -rf build/
$ mkdir build
$ cd build
$ CC=clang CXX=clang++ cmake -DCMAKE_BUILD_TYPE=DEBUG -DCMAKE_C_FLAGS_DEBUG="-g -O0" -DCMAKE_CXX_FLAGS_DEBUG="-g -O0" ..  # cmake with clang and debug/gdb symbols
...
$ make -j
...
[100%] Built target h264_vui_parameters_parser_unittest
```